backend/requirements.txt
# Backend Dependencies
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
pillow==10.0.1
pytesseract==0.3.10
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0

# Development
pytest==7.4.2
black==23.9.1
flake8==6.1.0
